added toggle navbar, frontend components
