--- component/book.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-import '../public/styles/main.css'
-
-
-function Book() {
-  return (
-    <>
-      <h1>Hello</h1>
-    </>
-  )
-}
-
-export default Book
