+++ src/App.jsx
@@ -2,12 +2,12 @@ import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
-import Header from '../component/header'
-import Home from '../component/home'
-import Services from '../component/services'
-import About from '../component/about'
-import Doctors from '../component/doctors'
-import Book from '../component/book'
+import Header from './component/header'
+import Home from './component/home'
+import Services from './component/services'
+import About from './component/about'
+
+
 
 
 function App() {
@@ -20,8 +20,8 @@ function App() {
       <Home /> 
       <Services /> 
       <About />
-      <Doctors />
-      <Book />
+     
+     
      
     </>
   )
